Use selectors more liberally
We are pulling in specific parts of the executedSearch response, as well
as the qtext for the preExecutedSearch (the one under construction by
the user). We are ALSO pulling in the entire preExecutedSearch, so we
can pass that to runSearch. This lets us stay decoupled from the shape
of that search query, though it does seem slightly redundant.