MUIR-9: Apply common eslint/prettier configs
Note that this time, I used

    "parserOptions": {
      "sourceType": "module",
      "ecmaVersion": "2018"
    },

in .eslintrc.json instead of babel-eslint, which I used in
muir-core-react-redux-containers, to get the spread operator and the
`import` syntax. I made that choice because babel is not used here. I
don't really like the "2018" ecmaVersion because I am only using spread,
and not the rest of 2018, but this is what the eslint core team is
recommending.